fix: KEEP-611 enforce the concurrency cap regardless of execution mode
The running-execution concurrency cap was only checked on the API route, so
schedule/block/event runs dispatched to k8s-job or in-process targets bypassed
it entirely. Extract the check into a server-only-free lib/workflow/concurrency
that both the route wrapper and the standalone executor share, and enforce it in
the executor before creating the execution row. On breach the executor throws so
the SQS message is redelivered after the visibility timeout (back-pressure)
rather than dropping the scheduled run.
